springConfiq divided into 3 classes
