+++ core/src/main/java/com/epam/esm/config/CoreSpringConfiq.java
@@ -1,0 +1,25 @@
+package com.epam.esm.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+
+@Configuration
+@EnableWebMvc
+@ComponentScan("com.epam.esm")
+public class CoreSpringConfiq {
+
+	@Bean
+	public Jackson2ObjectMapperBuilder objectMapperBuilder() {
+
+		Jackson2ObjectMapperBuilder builder = new Jackson2ObjectMapperBuilder();
+		builder.modules(new JavaTimeModule());
+
+		return builder;
+	}
+
+}
