--- core/src/main/java/com/epam/esm/util/parameter/ParameterConstant.java
@@ -1,16 +1,0 @@
-package com.epam.esm.util.parameter;
-
-public final class ParameterConstant {
-
-	public static final String TAG = "tag";
-	public static final String CERTIFICATE_NAME = "name";
-	public static final String DESCRIPTION = "description";
-	public static final String ORDER = "order";
-	public static final String DIRECTION = "direction";
-	
-	public static final String CREATION_DATE = "date";
-	public static final String DIRECTION_ASC = "asc";
-	public static final String DIRECTION_DESC = "desc";
-	
-
-}
