+++ dal/src/main/java/com/epam/esm/dal/config/DalSpringConfiq.java
@@ -1,28 +1,22 @@
-package com.epam.esm.rest.config;
+package com.epam.esm.dal.config;
 
 import javax.naming.NamingException;
 
 import org.apache.tomcat.dbcp.dbcp2.BasicDataSource;
-import org.modelmapper.ModelMapper;
-
 import org.springframework.beans.factory.annotation.Autowired;
-
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.env.Environment;
-import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
 @Configuration
 @EnableWebMvc
 @ComponentScan("com.epam.esm")
 @PropertySource("classpath:db.properties")
-public class SpringConfiq {
+public class DalSpringConfiq {
  
 	@Autowired
 	private Environment env;
@@ -45,19 +39,4 @@ public class SpringConfiq {
 		return new JdbcTemplate(dataSource());
 	}
 
-	@Bean
-	public Jackson2ObjectMapperBuilder objectMapperBuilder() {
-
-		Jackson2ObjectMapperBuilder builder = new Jackson2ObjectMapperBuilder();
-		builder.modules(new JavaTimeModule());
-
-		return builder;
-	}
-
-	@Bean
-	public ModelMapper modelMapper() {
-
-		ModelMapper mapper = new ModelMapper();
-		return mapper;
-	}
 }
